feat: primary button component
