+++ frontend/src/components/Sign/SignTitle.jsx
@@ -3,22 +3,21 @@ import { ColumnMiddleContainer } from 'components/common/Containers';
 import styled from 'styled-components';
 
 const Logo = styled.img`
-  width: 15rem;
-  height: 4rem;
+  width: 20rem;
 `;
 
 const P = styled.p`
   text-align: center;
-  font-family: 'NS-R';
-  font-size: 15px;
+  font-family: 'Pr-Regular';
+  font-size: 1rem;
 `;
 
 const SignTitle = () => {
   return (
     <ColumnMiddleContainer>
-      <Logo src="images/Logo.gif" alt="Logo" />
+      <Logo src="images/Logo.png" alt="Logo" />
       <br />
-      <P style={{ marginBottom: '5rem' }}>당신의 프로젝트, 여기서 구해보세요.</P>
+      <P style={{ marginBottom: '5rem' }}>당신의 능력, 티밍에서 펼쳐보세요.</P>
     </ColumnMiddleContainer>
   );
 };
